Add helper for determining .NET SDK version
Add a helper function to deduplicate the code to get the .NET SDK version in a repository and the line number it is defined on.
