+++ src/shared/github.ts
@@ -80,3 +80,45 @@ export async function getPull(
 export async function getWorkflowConfig(octokit: Api, context: Context): Promise<WorkflowConfig> {
   return JSON.parse(await getFileContents(octokit, context.repo.owner, context.repo.repo, '.github/workflow-config.json', context.sha));
 }
+
+export type RepositoryDotNetSdk = {
+  version: string;
+  line: number;
+};
+
+type GlobalJson = {
+  sdk: {
+    version: string;
+  };
+};
+
+export async function getDotNetSdk(octokit: Api, owner: string, repo: string, ref: string): Promise<RepositoryDotNetSdk | null> {
+  let globalJsonString: string;
+
+  try {
+    globalJsonString = await getFileContents(octokit, owner, repo, 'global.json', ref);
+  } catch (err) {
+    return null;
+  }
+
+  const globalJson: GlobalJson = JSON.parse(globalJsonString);
+  const sdkVersion = globalJson.sdk?.version;
+
+  let lineNumber = -1;
+
+  if (sdkVersion) {
+    const globalJsonLines = globalJsonString.split('\n');
+    for (let i = 0; i < globalJsonLines.length; i++) {
+      const line = globalJsonLines[i];
+      if (line.includes(sdkVersion)) {
+        lineNumber = i + 1;
+        break;
+      }
+    }
+  }
+
+  return {
+    version: sdkVersion,
+    line: lineNumber,
+  };
+}
